image_analysis_function/requirements.txt
functions-framework>=3.0
google-cloud-vision
google-cloud-storage
opencv-python-headless
numpy
